+++ src/main/java/tr/com/milia/resurgence/task/TaskResult.java
@@ -10,9 +10,9 @@ public abstract class TaskResult {
 
 	private final Player player;
 	private final Task task;
-	private final Map<Item, Integer> usedItems;
+	private final Map<Item, Long> usedItems;
 
-	protected TaskResult(Player player, Task task, Map<Item, Integer> usedItems) {
+	protected TaskResult(Player player, Task task, Map<Item, Long> usedItems) {
 		this.player = player;
 		this.task = task;
 		this.usedItems = usedItems;
@@ -30,7 +30,7 @@ public abstract class TaskResult {
 		return task;
 	}
 
-	public Map<Item, Integer> getUsedItems() {
+	public Map<Item, Long> getUsedItems() {
 		return Collections.unmodifiableMap(usedItems);
 	}
 }
